Drop all remaining 4e suffixes: one unsuffixed module per topic (#1188)
Now that every topic has a single version, rename the 4e-only modules (which had
no 3e counterpart) to plain topic names, and converge the notebook helper pair:
- deep_learning4e -> deep_learning, game_theory4e -> game_theory,
  making_simple_decisions4e -> making_simple_decisions, perception4e -> perception
  (modules + their tests renamed).
- notebook4e folded into notebook_utils (it was a superset: + plot_model_boundary);
  notebook4e deleted. (notebook_utils keeps the Jupyter-collision-safe name from #1294.)
- Repointed every import across modules, tests, notebooks, gui and the Sphinx docs;
  renamed game_theory4e.ipynb -> game_theory.ipynb.

No '*4e' modules remain. Full suite: 393 passed; docs build with 0 warnings.
