+++ aima/perception.py
