+++ tests/test_deep_learning.py
@@ -1,7 +1,7 @@
 import pytest
 from keras.datasets import imdb
 
-from aima.deep_learning4e import *
+from aima.deep_learning import *
 from aima.learning import DataSet, grade_learner, err_ratio
 
 random.seed("aima-python")
